requirements.txt
python-telegram-bot==20.7
googletrans==4.0.2